Fixes the set_asc_bit function when there is no 1.1V
This is the first set of broken code when there isn't 1.1V being supplied (1.1V -> VDDD). With the addition of the NOP it should get far enough into the program that you're able to to see some serial output. Additional fixes need to be added in order to complete the calibration code. The serial output is most likely at the wrong baud rate because calibration hasn't finished yet. Without the NOP, the "unsigned int position" variable being passed is set to zero. The position variable seems to be PUSHed onto the stack just fine but, when POPing, to restore it's value, it always sets it to zero. This can be better understood when looking at the assembly code.