1.1V fix
These changes fix the different modes in the frequency projects. Before, the 1.1V fixes only worked for the tx sweep. Now it should work for the tx/rx sweep and fixed mode.
The broken down print  statement in the optical.c file fixes the calibration output. Without the external 1.1V to VDDD the printed output had zeros for some of the variables it was trying to print. Calling any function with a lot of variables being passes is something that should be avoided when running the boards that need the 1.1V/VDDD tap fixes. This is because issues can pop up when there are a lot of loads back to back.